fix(api, auth): exporting AppRouter, oddly needed for better-auth
Potentially a side-effect type issue. If the AppRouter isn't exported,
the error we get is:

```
The inferred type of 'authClient' cannot be named without a reference to
'../../../../packages/auth/node_modules/better-auth/dist/shared/better-auth.BDR52Rf2'.
This is likely not portable. A type annotation is necessary.ts(2742)
```

Similar to:
- https://github.com/better-auth/better-auth/issues/4279
- https://github.com/better-auth/better-auth/issues/4054